--- azureclient/response.go
@@ -1,17 +1,0 @@
-package azureclient
-
-import (
-	"net/http"
-
-	"github.com/Azure/go-autorest/autorest"
-)
-
-// ResponseWasNotFound returns true if the response code from the Azure API
-// was a 404.
-func ResponseWasNotFound(resp autorest.Response) bool {
-	if resp.Response != nil && resp.StatusCode == http.StatusNotFound {
-		return true
-	}
-
-	return false
-}
