+++ service/controller/azureconfig/handler/dnsrecord/current.go
@@ -2,12 +2,13 @@ package dnsrecord
 
 import (
 	"context"
+	"net/http"
 
 	"github.com/Azure/azure-sdk-for-go/services/dns/mgmt/2018-05-01/dns"
+	"github.com/Azure/go-autorest/autorest"
 	providerv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/provider/v1alpha1"
 	"github.com/giantswarm/microerror"
 
-	"github.com/giantswarm/azure-operator/v5/azureclient"
 	"github.com/giantswarm/azure-operator/v5/service/controller/key"
 )
 
@@ -30,7 +31,7 @@ func (r *Resource) getCurrentState(ctx context.Context, obj providerv1alpha1.Azu
 
 	for i, record := range current {
 		resp, err := recordSetsClient.Get(ctx, record.ZoneRG, record.Zone, record.RelativeName, dns.NS)
-		if azureclient.ResponseWasNotFound(resp.Response) {
+		if responseWasNotFound(resp.Response) {
 			continue
 		} else if err != nil {
 			return nil, microerror.Mask(err)
@@ -45,4 +46,12 @@ func (r *Resource) getCurrentState(ctx context.Context, obj providerv1alpha1.Azu
 	}
 
 	return current, nil
+}
+
+func responseWasNotFound(resp autorest.Response) bool {
+	if resp.Response != nil && resp.StatusCode == http.StatusNotFound {
+		return true
+	}
+
+	return false
 }
